test(crdt): align consumers and navigation
